Extract getFileNamesFromFile to SRAUtil.cpp
